realizado acrescimento no formulário com o endereco e telefone

+++ app/repository/solicitacao-repository.js
@@ -5,8 +5,9 @@ const solicitacao = dataBase.sequelize.define('solicitacao', {
         type: dataBase.Sequelize.STRING
     },
     cliente_id: {
-        type: dataBase.Sequelize.BIGINT
-    }  
+        type: dataBase.Sequelize.INTEGER
+    }
+   
 }, { freezeTableName: true });
 
 module.exports = solicitacao
